Add script for bootstrapping echo test server
Added a new script that can be run once on an AWS EC2 server which can
set up a simple server that will echo back GET/POST requests sent to it.
This allows someone to easily set up a test server for running
integration tests that need to test connectivity or something similar.

To create an echo server, this script installs:
 - a simple python/flask server script in /home/ec2-user/
 - Gunicorn to provide the wsgi interface
 - Supervisord to keep the Gunicorn process running
 - Nginx used as a reverse proxy to the Gunicorn server