Add ListApiPaginator
The ListApiPaginator is a generic stream/iterator to use for any SDK
Client's List API calls.

Since the AWS SDK does not have a common interface for the clients, each
service is free to use whatever naming convention they like. Because of
this the ListApiPaginator needs to be given function objects for making
the required method calls (create Request object, make the API call and
get the list and token out of the return Result object)