Fix csv file reading bug in json2csv
Fixed bug where the json2csv script was throwing an error because the
DictReader was being closed before the file was read.

Also added logic to check if the input file's extension ends in '.csv'
in which case it will try to convert it to json (instead of the default
behaviour of converting to csv).