Add script for retrieving EOY Aus share prices
For US tax purposes, you have to declare the value of all foreign
finances at the end of the year. This script will help retrieve the
price of a given list of shares at the end of a given year, and retrieve
the AUD/USD exchange rate at that time.