ListApiCallIterator now returns empty list when there are no resources
Fixed a bug where the next() method would throw an exception because it
tried to return an item when in fact there were never any resources to
begin with.

Moved the logic to make the API calls into the hasNext() method, since
that is when we can actually know if there are more items to return.